+++ bin/render-build.sh
@@ -7,5 +7,3 @@ yarn install
 yarn build
 bundle exec rake assets:precompile
 bundle exec rake assets:clean
-bundle exec rake db:migrate:reset
-bundle exec rake db:seed