+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
 ## Unreleased
-###
+### Added
+- Setup admin model
 
 ## Version 1.0.0 (2023-10-15)
 
 ### New Features
-- Added company accounts
-- Added reverse auction
+- Implemented company authentication
+- Implemented reverse auction
 - Improved documentation
 - Added deployment per PR
+- Added github actions
+- Added render configuration
 
 ### Bug Fixes
 - Fixed javascript packaging